테스트 코드 3차 수정

+++ src/main/java/com/carenation/carenation_project/entity/Car.java
@@ -36,7 +36,7 @@ public class Car {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "CAR_ID")
-	private Long carId;
+	private Integer carId;
 
 	@Column(name = "MANUFACTURER", nullable = false)
 	private String manufacturer;
